+++ src/patterns/name-interpolation.js
@@ -18,15 +18,15 @@ module.exports = class NameInterpolationPattern extends BasePattern {
 
 			// Get the property
 			let property = _property;
-			if (property === 'aA') property = 'lowerCamel';
-			else if (property === 'AA') property = 'upperCamel';
-			else if (property === 'a') property = 'wordsLower';
-			else if (property === 'A') property = 'wordsUpper';
-			else if (property === 'a-a') property = 'hyphen';
-			else if (property === 'A-A') property = 'hyphenUpper';
-			else if (property === 'a_a') property = 'underscore';
-			else if (property === 'A_A') property = 'underscoreUpper';
-			else if (property === 'aa') property = 'oneWord';
+			if (property === 'aA') property = 'camel';
+			else if (property === 'AA') property = 'pascal';
+			else if (property === 'a') property = 'lower';
+			else if (property === 'A') property = 'capital';
+			else if (property === 'a-a') property = 'kebab';
+			else if (property === 'A-A') property = 'big';
+			else if (property === 'a_a') property = 'snake';
+			else if (property === 'A_A') property = 'constant';
+			else if (property === 'aa') property = 'compact';
 			else if (property === 'R') property = 'raw';
 			else throw new ParsingError(`[NameInterpolationPattern.execute] Unknown name property: ${property}`);